Extract gil management functions to separate header
